requirements.txt
numpy==1.19.1
scipy==1.4.1
matplotlib==3.3.2
opencv-python==4.4.0.46
torch_scatter
torch_geometric
torch_sparse
openmesh
scikit-learn
scikit-image==0.16.2
termcolor
chumpy==0.69
vctoolkit